+++ src/app/providers/AppProviders.tsx
@@ -1,19 +1,13 @@
-import { PropsWithChildren, Suspense } from 'react';
+import { PropsWithChildren } from 'react';
 import { QueryProvider } from './QueryProvider';
 import { Toaster, TooltipProvider } from 'shared/ui';
 import { FrontendObservability } from 'shared/config/';
-import { QueryParamsHandler } from './QueryParamsHandler';
 export function AppProviders({ children }: PropsWithChildren) {
   return (
     <>
       <FrontendObservability />
       <QueryProvider>
-        <TooltipProvider>
-          <Suspense>
-            <QueryParamsHandler />
-          </Suspense>
-          {children}
-        </TooltipProvider>
+        <TooltipProvider>{children}</TooltipProvider>
       </QueryProvider>
       <Toaster richColors />
     </>
