+++ src/entities/auth/api/queries.ts
@@ -1,18 +1,18 @@
 import { queryOptions } from '@tanstack/react-query';
 import { AuthHttp } from './http';
-import { authKeys } from '../model/const';
+import { authFabricKeys } from '../model/const';
 
 export class AuthQueries {
   static getOAuthProviders() {
     return queryOptions({
-      queryKey: authKeys.availableProviders(),
+      queryKey: authFabricKeys.availableProviders(),
       queryFn: async ({ signal }) => AuthHttp.oAuthProviders(signal),
       staleTime: 60_000 * 360 * 24,
     });
   }
   static getConnectedOAuthProviders() {
     return queryOptions({
-      queryKey: authKeys.connectedProviders(),
+      queryKey: authFabricKeys.connectedProviders(),
       queryFn: async ({ signal }) => AuthHttp.connectedOAuthProviders(signal),
       staleTime: 60_000 * 360 * 24,
     });
