+++ src/entities/auth/index.ts
@@ -3,4 +3,5 @@ export type * as TAuth from './model/types';
 export * as CAuth from './model/const';
 export { AuthHttp } from './api/http';
 export { AuthQueries } from './api/queries';
-export { type OAuthProviderMeta } from './model/const';
+export { authFabricKeys } from './model/const';
+export { OAUTH_PROVIDERS } from './config/oauth-providers';
