+++ src/features/auth/oauth-login/ui/OAuthLoginButtons.tsx
@@ -1,13 +1,12 @@
 'use client';
 import { Skeleton } from 'shared/ui';
-import { CAuth } from 'entities/auth';
 import { cn } from 'shared/lib/utils';
 import { useQuery } from '@tanstack/react-query';
 import { AuthQueries } from 'entities/auth';
-import { type TAuth } from 'entities/auth';
 import { routes } from 'shared/config';
-import { StartOauthParams } from '../model/types';
 import { OAuthButton } from './OAuthButton';
+import { type TAuth } from 'entities/auth';
+import { type StartOauthParams } from '../model/types';
 import { type Route } from 'next';
 
 export const getRoute = (provider: TAuth.OAuthProvider) => {
@@ -24,19 +23,10 @@ export function OAuthLoginButtons({ className }: { className?: string }) {
 
   return (
     <div className={cn('flex items-center justify-center gap-2', className)}>
-      {isLoading &&
-        Array.from({ length: CAuth.OAUTH_PROVIDERS_COUNT }, (_v, i) => (
-          <Skeleton className="size-8" key={i} />
-        ))}
+      {isLoading && Array.from({ length: 3 }, (_v, i) => <Skeleton className="size-8" key={i} />)}
 
       {data?.map((item) => {
-        return (
-          <OAuthButton
-            key={item.value}
-            provider={item.value}
-            href={getRoute(item.value) as Route}
-          />
-        );
+        return <OAuthButton key={item.value} data={item} href={getRoute(item.value) as Route} />;
       })}
     </div>
   );
