+++ src/features/handle-query-params/index.ts
@@ -1,0 +1,1 @@
+export { QueryParamsHandler } from './ui/QueryParamsHandler';
