+++ src/features/handle-query-params/ui/QueryParamsHandler.tsx
@@ -9,19 +9,22 @@ export function QueryParamsHandler() {
   const params = useSearchParams();
   const isShowToast = useRef(false);
   const router = useRouter();
+
   useEffect(() => {
     const success = params?.get('success');
     const message = params?.get('message');
 
     if (isShowToast.current || !success) return;
 
-    if (success === 'true' && message) {
-      toast.success(message);
-    } else if (success === 'false' && message) {
-      toast.error(message);
+    if (success === 'true') {
+      toast.success(message || 'Операция выполнена успешно');
+    } else if (success === 'false') {
+      toast.error(message || 'Произошла ошибка. Пожалуйста, попробуйте позже');
     }
+
     isShowToast.current = true;
     router.replace(location.pathname as Route);
   }, [params, router]);
+
   return null;
 }
