--- src/shared/ui/icon/types.ts
@@ -1,1 +1,0 @@
-export type IconProps = { className?: string };
