+++ src/shared/ui/index.ts
@@ -39,4 +39,3 @@ export * from './Select';
 export * from './Empty';
 export * from './ScrollArea';
 export * from './Kanban';
-export * from './icon';
